==> 1 Find high growth stocks with shift method.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import yfinance as yf
import xlwings as xw

high_growth_list = ["AAPL", "ACLS", 'AMD', 'AMPY', "AMZN", 'ARVN', 'ARWR', 'ATSG', 'AUDC','AYX', 
                    'BCRX', 'BPT','BRM', 'BRN', 'BTG', 'BTH', 'BTU', 'CALX', 'CLFD', 'CRM', 'CYP', 'DVN', 'DXCM',
                    'ELEF', 'ENPH', 'ETSY', 'EVER', 'EXCE', 'FCX', 'FIVN', 'FMC', 'FSM', 
                    'GLDG', 'GROY', 'GTN', 'IIPR', 'ILMN', 'IPA', 'KLIC',
                    'LAC',  'LRTNF', 'LSCC', 'LUGDF', 'MA', 'MAXR', 'MNST', 'MTEM', 'MU',
                    'NFLX', 'NINE', 'NRGU','NVDA', 'OESX', 'ON', 
                    'PBF','ROKU', 'SHOP', 'SI', 'SIRI',  'SMLR', 'SQ', 'STNG','SWKS',
                    'TELL', 'TH','TSLA', 'TTD',
                    'UEC', 'WYNN', 'ZM'
                   ]

lower_growth_list = ["ADBE", "AKBA", "AKRO", "AXP", "ALB", "ASC","ATNM", "ATVI", "AVAV", "AXP",
                     "BAC", "BAND", "CAG", "BRK-B","BBY", "BTU", "CMI","COIN", "COST", "EA",
                     "EBS", "EGIO", "EOG","ELV","EQX","EXCE","FE", "FSLR", "GLD", "GILD", "GOOG", 
                     "IVN", 'JNJ', 'JPM',"KEYS", "KLXE", "KNSL", 'KO',"LCR", 'LANC', "LLY", "LUGDF", "META",
                     "MSFT", 'MTSI', "MMM", "NEM", "NOW", "ORLY", "OTCM", "PH","PNK",
                     "QS" "RIG", "SBUX", 'SGEN','SPLS', 'TDOC', 'TSA', 'TSN', 'UAA','UI','V',"VEEV",
                     'VRTX', 'WDC','TRMD', "ZM"
                    ]

#Unfiltered for dates
growth_list_6mo = ['AKRO', 'COIN', 'COST', 'LLY', 'FE', 'BTH', 'BTU', 'EOG', 'BRK-B', 'RIG', 'DVN', 
                   'MTSI', 'QS', 'BRN', 'KEYS', 'ALB', 'CALX', 'NOW', 'ETSY', 'BAND', 'MAXR', 'EXCE',
                   'EQX', 'PH', 'LCR','EGIO', 'KNSL', 'AKBA', 'ARVN', 'EGIO', 'JPLS', 'ENPH', 'EVER', 
                   'OESX', 'ROKU', 'IIPR', 'AYX', 'AVAV', 'PNK', 'IPA', 'AMD', 'JNJ','FSM', 'NEM', 'BTG', 
                   'LRTNF', 'FIVN', 'CYP', 'OTCM', 'META', 'ATNM', 'EA', 'MTEM', 'IVN', 'NVDA', 'LSCC', 
                   'CRM', 'ACLS', 'GTN', 'CLFD', 'LUGDF', 'CAG', 'KLIC', 'SBUX', 'MSFT', 'BAC', 'AXP', 
                   'MMM','KO','BCRX','ELV', 'ADBE', 'CMI', 'ATSG', 'LAC', 'FCX', 'MU', 'DXCM', 'LANC', 
                   'ORLY', 'EBS', 'FSLR', 'MA', 'GLD', 'MNST', 'GOOG', 'AAPL', 'BRM', 'ARWR','ATVI' 
                   'ILMN', 'AMZN', 'AUDC', 'ON', 'GILD', 'NFLX']

#Filtered for 2018-2022
growth_list_6mo = ["AKBA", "ALB", "ARVN", "AVAV", "AYX", "BAND", "BRK-B", "BRN", "COIN",
                   "COST", "DVN", "EGIO", "ENPH", "EOG", "EQX", "ETSY", "EVER", "IIPR", "KEYS",
                   'KNSL', 'LCR', 'LLY', 'NOW', 'OESX','QS', 'ROKU', 'MAXR', 'MTSI', 'CALX',
                   'BTH', 'EXCE', 'FE', 'PH', 'RIG', 'BTU', 'JPLS']

ticker = "COST" 

df = yf.download(ticker, start = "2002-01-01", end = "2022-11-06")

type("Date")

df

df.info()

df.rename(columns = {"Adj Close": "Adj_Close"}, inplace = True)

df = df["Adj_Close"].copy().to_frame()

df.head()

df.Adj_Close.iloc[126]

df["Start_6mo"] = df.shift(periods = 126) #shift 6 months

df["Start_6mo"]

df["Start_2yr"] = df.Adj_Close.shift(periods = 504) #shift 6 months

df["Start_2yr"]

pd.set_option("display.max_rows", None)

df.isna().sum()

df["Abs_Change_2yr"] =df.Adj_Close.sub(df.Start_2yr)

df["Abs_Change_6mo"] = df.Adj_Close.sub(df.Start_6mo)

df["%_change_2yr"] =(df.Adj_Close / df.Start_2yr -1) * 100

df["%_change_2yr"]

df["%_change_6mo"] =(df.Adj_Close / df.Start_6mo -1) * 100

df["%_change_6mo"]

df.tail()

df["%_change_2yr"].max()

df["%_change_6mo"].max()

#These give the dates of the max 2yr & 6mo
result_2yr = df.iloc[df["%_change_2yr"].argmax()]
result_2yr

result_6mo = df.iloc[df["%_change_6mo"].argmax()]
result_6mo

#### Change the cell targets

path = "Model_Tickers2.xlsx"

wb = xw.Book(path)
wb

sheet = wb.sheets[0]
sheet

sheet.range("A150").value = ticker

sheet.range("A150").value = result_2yr

sheet.range("A150").value = result_6mo
