==> 3 Features on Each Stock.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import yfinance as yf
import xlwings as xw

### Feature characteristics:
#### Positive Earnings Revisions
#### Positive Earnings Surprises
     Both of these not easy to get in retrospect - see if yfinance has anything like it.
#### Sales Growth
     Year-over-year increase of 20% or more - need revenue this year and last year
#### Expanding Operating Margins
     Expanding year over year and quarter over quarter - need previous quarterly & yearly
#### Free Cash Flow
     FCF compared to market capitalization - need FCF and market cap -CF statement CFFO-Capex
#### Earnings Growth
     Growing consistently y-o-y 10% or more in last quarter
#### Positive Earnings Momentum
     Earnings growing at an increasing rate - rate of growth of earnings this quarter compared to last
#### Return on Equity
    Starts high and moves higher


growth_list_6mo = ["AKBA", "ALB", "ARVN", "AYX", 'BRK-B',"BAND", "BRN", "COIN", 
                    "COST", "DVN", "EGIO", "ENPH", "EOG", "EQX", "ETSY", "EVER", "IIPR", "KEYS", 
                    'KNSL', 'LLY', 'NOW', 'OESX','QS', 'ROKU', 'MAXR', 'MTSI', 'CALX',
                    'FE', 'PH', 'RIG', 'BTU']

ticker = "COST"  

df = yf.download(ticker, start = "2002-01-01", end = "2022-11-25")

type("Date")

ticker

df

df.info()

df.rename(columns = {"Adj Close": "Adj_Close"}, inplace = True)

df = df["Adj_Close"].copy().to_frame()

df.head()

df.Adj_Close.iloc[126]

df["6mo"] = df.shift(periods = 126) #shift 6 months

df["6mo"]

df["Start"] = df.Adj_Close.shift(periods = 504) #shift 6 months

df["Start"]

df

df["Abs_Change_2yr"] =df.Adj_Close.sub(df.Start)

df["Abs_Change_6mo"] = df["6mo"].sub(df.Start)

df

df["%_change_2yr"] =(df.Adj_Close / df.Start -1) * 100

df["%_change_2yr"]

df["%_change_6mo"] =((df.Adj_Close / df["6mo"]) -1) * 100

df["%_change_6mo"]

df.tail()

df["%_change_2yr"].max()

Pct_change_6mo = df["%_change_6mo"].max()
Pct_change_6mo

df.iloc[df["%_change_2yr"].argmax()]

df.iloc[df["%_change_6mo"].argmax()]

### Get fundamental data

#First make a ticker object 2 steps
# ticker = "AKBA"  #can add more
stock = yf.Ticker(ticker)

df_prices = yf.download(ticker, start = "2018-01-01", end = "2022-11-25") #??

#Access the info keys
stock_info = stock.info
#stock_info.keys()

stock_info["symbol"]

#access data using the keys.  It doesn't get all of them
sector = stock_info["sector"]
sector

stock.info["marketCap"]

income_stmt = stock.get_financials()
income_stmt = income_stmt.astype("float")
#stock_pnl.info()
income_stmt

cashflow_stmt = stock.get_cashflow()
cashflow_stmt = cashflow_stmt.astype("float")
# cashflow_stmt.info()
cashflow_stmt

#roku.balancesheet
balance_sheet = stock.get_balancesheet()
balance_sheet = balance_sheet.astype("float")
#balance_sheet.info()
balance_sheet

years = income_stmt.columns
years

### Set by date of interest

#Sales growth (This is displayed as %)
#Current year
revenue_current = income_stmt[years[0]]["Total Revenue"] 
#Past year
revenue_last = income_stmt[years[1]]["Total Revenue"]
sales_growth= ((revenue_current / revenue_last) -1) *100
sales_growth

#Expanding operating margin (displayed as %)
#Current year
operating_margin_current = income_stmt[years[0]]["Operating Income"] - income_stmt[years[0]]["Total Operating Expenses"]
#Past year
operating_margin_last = income_stmt[years[1]]["Operating Income"] - income_stmt[years[1]]["Total Operating Expenses"]
operating_margin_change = ((operating_margin_current / operating_margin_last) -1) *100
operating_margin_change

#Free cash flow
#Past year for both
cashflow_oper_act = cashflow_stmt[years[1]]["Total Cash From Operating Activities"]
capex = cashflow_stmt[years[1]]["Capital Expenditures"]
free_cash_flow = cashflow_oper_act + capex #CAPEX is in negative numbers
free_cash_flow

#market cap = shares outstanding * price of shares #Use current & date of interest
#Past year
shares_out = balance_sheet[years[2]]["Common Stock"]
#Date of interest
price = df.loc["2021-09-03", "Adj_Close"]
marketcap = shares_out * price
shares_out
marketcap

#free cash flow yield = fcf / market cap
fcf_yield = free_cash_flow/marketcap
fcf_yield

#Additional calculation of fcf_yield
#Free cash flow per share/price per share
fcf_share = free_cash_flow/shares_out
price_per_share = df.iloc[df["%_change_6mo"].argmax()][0]
# price_per_share
fcf_yield = fcf_share/price_per_share
fcf_yield

#Expanding earnings (displayed as %)
#current year
earnings_current = income_stmt[years[0]]["Operating Income"]
#Past year
earnings_last = income_stmt[years[1]]["Operating Income"]
earnings_change = ((earnings_current / earnings_last) -1) *100
earnings_change

#Debt ratio = total debt /total assets (displayed as %)
#Past year
total_debt = balance_sheet[years[1]]["Long Term Debt"]+ balance_sheet[years[1]]["Short Long Term Debt"] 
#Past year
total_assets = balance_sheet[years[1]]["Total Assets"]
debt_ratio = total_debt / total_assets
debt_ratio


#Return on equity = Net income / shareholders' equity
#Past year for both
net_income = income_stmt[years[1]]["Net Income"] 
shareholders_equity = balance_sheet[years[1]]["Total Stockholder Equity"]
return_on_equity = net_income / shareholders_equity
return_on_equity = return_on_equity * 100
return_on_equity


#Return on invested capital = NOPAT / Invested capital, effective tax rate = tax expense / net income
#Past year for both
effective_tax_rate = income_stmt[years[1]]["Income Tax Expense"] / income_stmt[years[1]]["Net Income"]
operating_profit = income_stmt[years[1]]["Operating Income"] - income_stmt[years[1]]["Total Operating Expenses"]
#operating_profit
NOPAT = operating_profit * (1 - effective_tax_rate)
#Past year
invested_capital = balance_sheet[years[1]]["Long Term Debt"]+ balance_sheet[years[1]]["Short Long Term Debt"]
+ balance_sheet[years[2]]["Total Stockholder Equity"]
ROIC = NOPAT / invested_capital
ROIC

total_assets = balance_sheet[years[1]]["Total Assets"]
total_assets

###  Insert this into Excel via XLwings 

path = "Model_Tickers.xlsx"

wb = xw.Book(path)
wb

sheet2 = wb.sheets[2]
sheet2

sheet2.range("A34").value = ticker

sheet2.range("B34").value = #date of interest

sheet2.range("C34").value = Pct_change_6mo

sheet2.range("D34").value = sector

sheet2.range("E34").value = sales_growth

sheet2.range("F34").value = operating_margin_change

sheet2.range("G34").value = free_cash_flow

sheet2.range("H34").value = marketcap

sheet2.range("I34").value = fcf_yield

sheet2.range("J34").value = earnings_change

sheet2.range("K34").value = debt_ratio

sheet2.range("L31").value = return_on_equity

sheet2.range("M34").value = ROIC

sheet2.range("N34").value = total_assets
