==> 4  Add Features to Stock List.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import yfinance as yf
import xlwings as xw

df = pd.read_excel("Model_Tickers2.xlsx", index_col = 0).T
df 

df.head()

df_6mo = df.drop(["Abs_Change_2yr", "Start_6mo-R","Abs_Change_6mo", "%_change_6mo-R", "Start_2yr-R", "%_change_2yr-R"], axis = 1, inplace = True)

df_6mo

pd.set_option('display.max_rows', None)

df.sort_values(by = "Start_Date_6mo", ascending = False)

df.shape

date_col = pd.to_datetime(df["Start_Date_6mo"], format= '%Y-%m-%d')

filtered_date_col = date_col.loc[(date_col >= '2018') & (date_col <= '2022')]
filtered_date_col

filtered_date_col.value_counts().sum()

df_6mo = df[(df["Start_Date_6mo"] >= '2018') & (df["Start_Date_6mo"] <= '2022')].drop(columns = ["Start_Date_2yr", "Adj_Close_2yr", "%_change_2yr", "Start_2yr"]).copy()

df_6mo

df_6mo.shape

df_6mo["Target"] = df_6mo["%_change_6mo"].apply(lambda x: 1 if x >= 150 else 0)

df_6mo

df_6mo.Target.value_counts()

features = pd.read_excel("Model_Tickers.xlsx", sheet_name="Sheet2", header=0, index_col = "Ticker")

features

features.drop(["Total_Assets", "ROIC"], axis = 1, inplace = True)
features

len(features)

len(df_6mo) + len(features)

df_6mo_features = pd.merge(df_6mo, features, left_index = True, right_index = True)
df_6mo_features

len(df_6mo_features)

df_6mo_features.to_csv("df_6mo_features.csv")
