==> 5 Perform Analysis with Support Vector Machine Model.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

df = pd.read_csv("df_6mo_features.csv", index_col = 0)

df.index.rename("Ticker", inplace = True)

df.head()

df.Target.unique()

df.Target.value_counts()

sns.countplot(x = "Target", data=df)

df = df.drop(['Start_Date_6mo', 'Adj_Close_6mo', 'Start_6mo', '%_change_6mo', 'Date_of_interest'], axis = 1)

df["Exceeds_150%"] = (df["Target"] == 1)

df.corr()["Exceeds_150%"]

df.corr()["Exceeds_150%"][:][2:-1].sort_values().plot(kind="bar", cmap = "rainbow_r")

plt.figure(figsize = (12,6))
sns.scatterplot(x = "Ticker", y = "Pct_change_6mo", data = df, s = 140, hue = "Earnings Change", palette = 'rocket')
plt.xticks(rotation = 45)
plt.show()

df

### Analyze with Support Vector Machine Model

X  = df.drop(["Target", "Pct_change_6mo", "Industry", "Exceeds_150%"],axis = 1)
y = df["Target"]

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = .2, random_state = 101)

X_train.head()

from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()

scaled_X_train = scaler.fit_transform(X_train)
scaled_X_test = scaler.transform(X_test)

from sklearn.svm import SVC

svc = SVC(class_weight = "balanced")

from sklearn.model_selection import GridSearchCV

param_grid = {'C': [0.001, 0.01, 0.1, 0.5, 1], 'gamma': ['scale', 'auto']}
grid = GridSearchCV(svc, param_grid)

grid.fit(scaled_X_train, y_train)

grid.best_params_

from sklearn.metrics import confusion_matrix, classification_report

grid_pred = grid.predict(scaled_X_test)
grid_pred

confusion_matrix(y_test, grid_pred)

print(classification_report(y_test, grid_pred, labels = np.unique(grid_pred)))

df.columns[3:11]

new_stock = [[-11.848920,-28.302362, -264096000, 2.670000e+03, -98912.356724, -300.174127, 0.098295,0.708433]]

new_stock = np.array(new_stock).reshape(1,-1)

grid.predict(new_stock)
